Separate shared selectors and tests
