+++ browser-tests/berth.ts
@@ -2,26 +2,27 @@ import { berth } from './selectors/berth';
 import { navbar } from './selectors/navbar';
 import { isBerthsPage } from './utils/page';
 import { envUrl } from './utils/settings';
+import { applicantInformation, overview, yourSelection } from './utils/sharedTests';
 
 const testData = {
+  address: 'Testiosoite 1',
+  boatDraught: '0.21',
+  boatLength: '4.4',
+  boatModel: 'Terhi 440',
+  boatName: 'Testivene',
+  boatRegistrationNumber: 'P12345',
   boatType: 'Soutuvene',
   boatTypeIndex: '2',
-  boatWidth: '1.75',
-  boatLength: '4.4',
-  harbor1: 'Lähteelä retkisatama',
-  harbor2: 'Strömsinlahden venesatama',
-  boatRegistrationNumber: 'P12345',
-  boatDraught: '0.21',
   boatWeight: '150',
-  boatName: 'Testivene',
-  boatModel: 'Terhi 440',
+  boatWidth: '1.75',
+  choice1: 'Lähteelä retkisatama',
+  choice2: 'Strömsinlahden venesatama',
+  emailAddress: 'test@example.com',
   firstName: 'Matti',
   lastName: 'Meikäläinen',
-  address: 'Testiosoite 1',
-  postalCode: '02100',
   municipality: 'Espoo',
   phoneNumber: '+358 50 123 4567',
-  emailAddress: 'test@example.com',
+  postalCode: '02100',
 };
 
 fixture('Berth').page(envUrl());
@@ -31,10 +32,21 @@ test('New berth application, registered boat, private customer', async (t) => {
   await isBerthsPage();
 
   await selectHarbors(t);
-  await yourSelection(t);
+  await yourSelection(t, testData);
   await boatInformation(t);
-  await applicantInformation(t);
-  await overview(t);
+  await applicantInformation(t, testData);
+
+  const expectedBoatInfo = [
+    `Nimi: ${testData.boatName}`,
+    `Rekisterinumero: ${testData.boatRegistrationNumber}`,
+    `Tyyppi: ${testData.boatType}`,
+    `Malli: ${testData.boatModel}`,
+    `Leveys: ${testData.boatWidth}m`,
+    `Pituus: ${testData.boatLength}m`,
+    `Syväys: ${testData.boatDraught}m`,
+    `Paino: ${testData.boatWeight}`,
+  ].join('\n');
+  await overview(t, testData, expectedBoatInfo);
 });
 
 const selectHarbors = async (t: TestController) => {
@@ -61,24 +73,8 @@ const selectHarbors = async (t: TestController) => {
 
   await t
     .click(harborListTab)
-    .click(getSelectButtonForHarbor(testData.harbor1))
-    .click(getSelectButtonForHarbor(testData.harbor2));
-
-  await t.click(nextButton);
-};
-
-const yourSelection = async (t: TestController) => {
-  const { heading, getHarborHeading, getUpButtonForHeading, nextButton } = berth.yourSelection;
-
-  await t.expect(heading.exists).ok();
-
-  await t.click(getUpButtonForHeading(getHarborHeading(`2. ${testData.harbor2}`))).wait(500);
-
-  await t
-    .expect(getHarborHeading(`1. ${testData.harbor2}`).exists)
-    .ok()
-    .expect(getHarborHeading(`2. ${testData.harbor1}`).exists)
-    .ok();
+    .click(getSelectButtonForHarbor(testData.choice1))
+    .click(getSelectButtonForHarbor(testData.choice2));
 
   await t.click(nextButton);
 };
@@ -104,76 +100,4 @@ const boatInformation = async (t: TestController) => {
     .typeText(boatModel, testData.boatModel);
 
   await t.click(nextButton);
-};
-
-const applicantInformation = async (t: TestController) => {
-  const {
-    heading,
-    firstName,
-    lastName,
-    address,
-    postalCode,
-    municipalitySelect,
-    phoneNumber,
-    emailAddress,
-    nextButton,
-  } = berth.applicantInformation;
-
-  await t.expect(heading.exists).ok();
-
-  await t
-    .typeText(firstName, testData.firstName)
-    .typeText(lastName, testData.lastName)
-    .typeText(address, testData.address)
-    .typeText(postalCode, testData.postalCode)
-    .click(municipalitySelect)
-    .click(municipalitySelect.find('option').withText(testData.municipality))
-    .typeText(phoneNumber, testData.phoneNumber)
-    .typeText(emailAddress, testData.emailAddress);
-
-  await t.click(nextButton);
-};
-
-const overview = async (t: TestController) => {
-  const { heading, textInOverview, getLabelValuePairs } = berth.overview;
-
-  await t.expect(heading.exists).ok();
-
-  // Boat information
-  const labelValuePairs = await getLabelValuePairs();
-  const expectedContent = [
-    `Nimi: ${testData.boatName}`,
-    `Rekisterinumero: ${testData.boatRegistrationNumber}`,
-    `Tyyppi: ${testData.boatType}`,
-    `Malli: ${testData.boatModel}`,
-    `Leveys: ${testData.boatWidth}m`,
-    `Pituus: ${testData.boatLength}m`,
-    `Syväys: ${testData.boatDraught}m`,
-    `Paino: ${testData.boatWeight}`,
-  ].join('\n');
-  await t.expect(labelValuePairs).eql(expectedContent);
-
-  // Chosen harbors
-  await t
-    .expect(textInOverview(`1. ${testData.harbor2}`).exists)
-    .ok()
-    .expect(textInOverview(`2. ${testData.harbor1}`).exists)
-    .ok();
-
-  // Applicant
-  await t
-    .expect(textInOverview(testData.firstName).exists)
-    .ok()
-    .expect(textInOverview(testData.lastName).exists)
-    .ok()
-    .expect(textInOverview(testData.emailAddress).exists)
-    .ok()
-    .expect(textInOverview(testData.phoneNumber).exists)
-    .ok()
-    .expect(textInOverview(testData.address).exists)
-    .ok()
-    .expect(textInOverview(testData.postalCode).exists)
-    .ok()
-    .expect(textInOverview(testData.municipality).exists)
-    .ok();
 };
